# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [Skincare, Makeup, Haircare, Bodycare, Fragrance]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /matches:
    get:
      operationId: listMatches
      tags: [products]
      summary: List all product matches
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [Skincare, Makeup, Haircare, Bodycare, Fragrance]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of product match pairs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductMatch"
  /matches/{matchId}:
    get:
      operationId: getMatch
      tags: [products]
      summary: Get a specific product match by match ID
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product match detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductMatch"
        "404":
          description: Match not found
...
  /products/luxury:
    get:
      operationId: listLuxuryProducts
      tags: [products]
      summary: List all luxury products (catalog)
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [Skincare, Makeup, Haircare, Bodycare, Fragrance]
      responses:
        "200":
          description: List of luxury products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LuxuryProduct"
  /products/similar:
    get:
      operationId: listSimilarLuxuryProducts
      tags: [products]
      summary: Find luxury products similar to a query (fuzzy match on brand+name)
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
      responses:
        "200":
          description: List of similar luxury products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimilarLuxuryProduct"
...
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: List all products (admin)
      parameters:
        - name: x-admin-password
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: All products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Create a new product (admin)
      parameters:
        - name: x-admin-password
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
  /admin/products/{id}:
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Delete a product (admin)
      parameters:
        - name: x-admin-password
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        price:
          type: number
          format: float
        imageUrl:
          type: string
        affiliateLink:
          type: string
        category:
          type: string
          enum: [Skincare, Makeup, Haircare, Bodycare, Fragrance]
        type:
          type: string
          enum: [Luxury, Dupe]
        matchId:
          type: integer
        matchScore:
          type: integer
        formato:
          type: ["number", "null"]
          format: float
        unitaMisura:
          type: ["string", "null"]
        pricePerUnit:
          type: ["number", "null"]
          format: float
        dupeTier:
          type: ["string", "null"]
          enum: [budget, mid-range, premium-dupe, null]
        luxuryGroupId:
          type: ["integer", "null"]
        aiMatchReason:
          type: ["string", "null"]
      required:
        - id
        - name
        - brand
        - price
        - imageUrl
        - affiliateLink
        - category
        - type
        - matchId
        - matchScore
        - formato
        - unitaMisura
        - pricePerUnit
        - dupeTier
        - luxuryGroupId
        - aiMatchReason
    ProductMatch:
      type: object
      properties:
        matchId:
          type: integer
        category:
          type: string
        luxury:
          $ref: "#/components/schemas/Product"
        dupe:
          $ref: "#/components/schemas/Product"
        matchScore:
          type: integer
        priceDifference:
          type: number
          format: float
        savingsPercent:
          type: number
          format: float
      required:
        - matchId
        - category
        - luxury
        - dupe
        - matchScore
        - priceDifference
        - savingsPercent
...
    LuxuryProduct:
      type: object
      properties:
        id:
          type: integer
        brand:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [Skincare, Makeup, Haircare, Bodycare, Fragrance]
        imageUrl:
          type: string
        price:
          type: ["number", "null"]
          format: float
        isAnalyzed:
          type: boolean
      required:
        - id
        - brand
        - name
        - category
        - imageUrl
        - price
        - isAnalyzed
...
    SimilarLuxuryProduct:
      type: object
      properties:
        id:
          type: integer
        brand:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [Skincare, Makeup, Haircare, Bodycare, Fragrance]
        imageUrl:
          type: string
        price:
          type: ["number", "null"]
          format: float
        isAnalyzed:
          type: boolean
      required:
        - id
        - brand
        - name
        - category
        - imageUrl
        - price
        - isAnalyzed
...
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        price:
          type: number
          format: float
        imageUrl:
          type: string
        affiliateLink:
          type: string
        category:
          type: string
          enum: [Skincare, Makeup, Haircare, Bodycare, Fragrance]
        type:
          type: string
          enum: [Luxury, Dupe]
        matchId:
          type: integer
        matchScore:
          type: integer
        formato:
          type: ["number", "null"]
          format: float
        unitaMisura:
          type: ["string", "null"]
      required:
        - name
        - brand
        - price
        - imageUrl
        - affiliateLink
        - category
        - type
        - matchId
        - matchScore