# lib/api-zod/src/generated/types/similarLuxuryProduct.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SimilarLuxuryProductCategory } from "./similarLuxuryProductCategory";

export interface SimilarLuxuryProduct {
  id: number;
  brand: string;
  name: string;
  category: SimilarLuxuryProductCategory;
  imageUrl: string;
  /** @nullable */
  price: number | null;
  isAnalyzed: boolean;
}